Testing infrastructure fixes
New, undocumented-on-purpose -mocktime=timestamp command-line
argument to startup with mocktime set. Needed because
time-related blockchain sanity checks are done on startup, before a
test has a chance to make a setmocktime RPC call.

And changed the setmocktime RPC call so calling it will not result in
currently connected peers being disconnected due to inactivity timeouts.

DOES NOT fix a -regest bug, unlike the version of this commit for 0.11.2:

Blocks generated in -regtest mode would pay to a key that was left
in the keypool, which can cause subtle bugs. In particular, I expected
every block generated to be unique, but got duplicate blocks between
calls to setgenerate.

This was Bitcoin Core issue 6268. It was fixed in PR #5994, so is
unnecessary in 0.12 and later.

Conflicts:
	src/init.cpp
	src/rpcmisc.cpp