Two megabyte fork after miner vote and grace period
Activation code for a two megabyte block size increase.

Activation condition:

750 of 1000 blocks with version bit 0x10000000 set, then a four-week
(28-day) grace period, after which miners may create 2MB
blocks.

You must run with the -vote2mb switch/option to produce blocks
with the voting bit set, but if you run this code your node
will follow the 2MB branch of the chain.

This is the 0.12 version. It has a few differences from the 0.11.2
version. These are mostly small bugs that were corrected in later commits
in the 0.11.2 branch which Jonathan decided would be better to not commit
at all in 0.12:

- In init.cpp, the help message for -blockmaxsize correctly lists the
  default as being for the max size possible, instead of 750kB.
- IsSuperMajority()'s default value for useBitMask is false.
- In ForkBits(), sizeForkTime.load() < nTime has the < replaced with
  a >.
- It contains different code in txdb.cpp due to the new leveldb code
  in 0.12.
- The default block size for qa/rpc-tests/prioritise_transaction.py is
  set explicitly to 750,000 in order to prevent a 0.12-only failure.

Note that qa/rpc-tests/bigblocks.py is incomplete in this commit and will
fail.